refactor(agents): split episodic.py and server.py (v0.2 release prep A-3)
Extract SQL query helpers and data model from agents/memory/episodic.py
into agents/memory/episodic_queries.py, and persona-path agent loading
and lifecycle helpers from agents/server.py into agents/server_persona.py.

Both source files are now ≤ 500 lines; no public API surface changed.
Update test mock targets from agents.server to agents.server_persona
for the relocated create_provider import. Mark PR 2 (#96) merged and
PR 3 open in docs/v0.2-release-prep-plan.md; add PRs #94–#96 to ROADMAP.

File sizes after split:
- agents/memory/episodic.py:        669 → 463 lines
- agents/memory/episodic_queries.py:         0 → 275 lines (new)
- agents/server.py:                 629 → 454 lines
- agents/server_persona.py:           0 → 228 lines (new)
