+++ FILEMAP.md
@@ -1,14 +1,14 @@
 # Repository File Map
 
-> Auto-generated by `scripts/generate_filemap.py` — 2026-04-17  
-> **229** tracked files
+> Auto-generated by `scripts/generate_filemap.py` — 2026-04-18  
+> **231** tracked files
 
 ## Summary
 
 | Top-level directory | Files |
 |---------------------|-------|
 | `internal/` | 52 |
-| `agents/` | 39 |
+| `agents/` | 41 |
 | `docs/` | 34 |
 | `tests/` | 30 |
 | `(root)` | 20 |
@@ -26,7 +26,7 @@
 
 | Extension | Count |
 |-----------|-------|
-| `.py` | 78 |
+| `.py` | 80 |
 | `.md` | 49 |
 | `.go` | 48 |
 | `.yaml` | 22 |
@@ -93,6 +93,7 @@
 │   ├── memory/
 │   │   ├── __init__.py
 │   │   ├── episodic.py
+│   │   ├── episodic_queries.py
 │   │   ├── migrations.py
 │   │   ├── notes.py
 │   │   ├── relationship.py
@@ -107,6 +108,7 @@
 │   ├── persona_types.py
 │   ├── pyproject.toml
 │   ├── server.py
+│   ├── server_persona.py
 │   ├── sub_agents/
 │   │   ├── __init__.py
 │   │   ├── spawner.py
